Rename TypeChecker and TypeCheckBuilder
