+++ python-frontend/src/test/java/org/sonar/python/types/v2/TypeCheckerContextTest.java
@@ -36,9 +36,9 @@ import static org.sonar.python.types.v2.TypesTestUtils.PROJECT_LEVEL_TYPE_TABLE;
 import static org.sonar.python.types.v2.TypesTestUtils.parseAndInferTypes;
 
 
-class TypeCheckerTest {
+class TypeCheckerContextTest {
 
-  TypeChecker typeChecker = new TypeChecker(PROJECT_LEVEL_TYPE_TABLE);
+  TypeCheckerContext typeCheckerContext = new TypeCheckerContext(PROJECT_LEVEL_TYPE_TABLE);
 
   @Test
   void isBuiltinWithNameTest() {
@@ -47,18 +47,18 @@ class TypeCheckerTest {
     ObjectType intLiteralType = (ObjectType) intLiteral.typeV2();
     assertThat(intLiteralType.unwrappedType()).isEqualTo(INT_TYPE);
 
-    assertThat(typeChecker.typeCheckBuilder().isBuiltinWithName("int").check(intLiteralType)).isEqualTo(TriBool.TRUE);
-    assertThat(typeChecker.typeCheckBuilder().isBuiltinWithName("str").check(intLiteralType)).isEqualTo(TriBool.FALSE);
-    assertThat(typeChecker.typeCheckBuilder().isBuiltinWithName("unknown").check(intLiteralType)).isEqualTo(TriBool.UNKNOWN);
+    assertThat(typeCheckerContext.typeChecker().isBuiltinWithName("int").check(intLiteralType)).isEqualTo(TriBool.TRUE);
+    assertThat(typeCheckerContext.typeChecker().isBuiltinWithName("str").check(intLiteralType)).isEqualTo(TriBool.FALSE);
+    assertThat(typeCheckerContext.typeChecker().isBuiltinWithName("unknown").check(intLiteralType)).isEqualTo(TriBool.UNKNOWN);
 
     fileInput = parseAndInferTypes("foo()");
     CallExpression callExpression = (CallExpression) TreeUtils.firstChild(fileInput, t -> t.is(Tree.Kind.CALL_EXPR)).get();
     PythonType callExpressionType = callExpression.typeV2();
     assertThat(callExpressionType).isEqualTo(PythonType.UNKNOWN);
-    assertThat(typeChecker.typeCheckBuilder().isBuiltinWithName("int").check(callExpressionType)).isEqualTo(TriBool.UNKNOWN);
-    assertThat(typeChecker.typeCheckBuilder().isBuiltinWithName("unknown").check(callExpressionType)).isEqualTo(TriBool.UNKNOWN);
+    assertThat(typeCheckerContext.typeChecker().isBuiltinWithName("int").check(callExpressionType)).isEqualTo(TriBool.UNKNOWN);
+    assertThat(typeCheckerContext.typeChecker().isBuiltinWithName("unknown").check(callExpressionType)).isEqualTo(TriBool.UNKNOWN);
 
-    assertThat(typeChecker.typeCheckBuilder().isBuiltinWithName("int").check(INT_TYPE)).isEqualTo(TriBool.TRUE);
+    assertThat(typeCheckerContext.typeChecker().isBuiltinWithName("int").check(INT_TYPE)).isEqualTo(TriBool.TRUE);
   }
 
 
@@ -73,8 +73,8 @@ class TypeCheckerTest {
     );
     ObjectType objectType = (ObjectType) ((ExpressionStatement) fileInput.statements().statements().get(2)).expressions().get(0).typeV2();
 
-    assertThat(typeChecker.typeCheckBuilder().hasMember("foo").check(objectType)).isEqualTo(TriBool.FALSE);
-    assertThat(typeChecker.typeCheckBuilder().instancesHaveMember("foo").check(objectType)).isEqualTo(TriBool.FALSE);
+    assertThat(typeCheckerContext.typeChecker().hasMember("foo").check(objectType)).isEqualTo(TriBool.FALSE);
+    assertThat(typeCheckerContext.typeChecker().instancesHaveMember("foo").check(objectType)).isEqualTo(TriBool.FALSE);
 
     fileInput = parseAndInferTypes("""
       class A:
@@ -84,17 +84,17 @@ class TypeCheckerTest {
       """
     );
     objectType = (ObjectType) ((ExpressionStatement) fileInput.statements().statements().get(2)).expressions().get(0).typeV2();
-    assertThat(typeChecker.typeCheckBuilder().hasMember("foo").check(objectType)).isEqualTo(TriBool.TRUE);
-    assertThat(typeChecker.typeCheckBuilder().instancesHaveMember("foo").check(objectType)).isEqualTo(TriBool.FALSE);
+    assertThat(typeCheckerContext.typeChecker().hasMember("foo").check(objectType)).isEqualTo(TriBool.TRUE);
+    assertThat(typeCheckerContext.typeChecker().instancesHaveMember("foo").check(objectType)).isEqualTo(TriBool.FALSE);
   }
 
   @Test
   void classTypeHasMemberTest() {
     ClassType classType = classType("class C: ...");
 
-    assertThat(typeChecker.typeCheckBuilder().hasMember("__call__").check(classType)).isEqualTo(TriBool.TRUE);
-    assertThat(typeChecker.typeCheckBuilder().hasMember("unknown").check(classType)).isEqualTo(TriBool.UNKNOWN);
-    assertThat(typeChecker.typeCheckBuilder().instancesHaveMember("__call__").check(classType)).isEqualTo(TriBool.FALSE);
+    assertThat(typeCheckerContext.typeChecker().hasMember("__call__").check(classType)).isEqualTo(TriBool.TRUE);
+    assertThat(typeCheckerContext.typeChecker().hasMember("unknown").check(classType)).isEqualTo(TriBool.UNKNOWN);
+    assertThat(typeCheckerContext.typeChecker().instancesHaveMember("__call__").check(classType)).isEqualTo(TriBool.FALSE);
   }
 
   @Test
@@ -104,16 +104,16 @@ class TypeCheckerTest {
     ObjectType intLiteralType = (ObjectType) intLiteral.typeV2();
     assertThat(intLiteralType.unwrappedType()).isEqualTo(INT_TYPE);
 
-    assertThat(typeChecker.typeCheckBuilder().isBuiltinWithName("str").or(
-      typeChecker.typeCheckBuilder().isBuiltinWithName("int")
+    assertThat(typeCheckerContext.typeChecker().isBuiltinWithName("str").or(
+      typeCheckerContext.typeChecker().isBuiltinWithName("int")
     ).check(intLiteralType)).isEqualTo(TriBool.TRUE);
 
-    assertThat(typeChecker.typeCheckBuilder().isBuiltinWithName("int").or(
-      typeChecker.typeCheckBuilder().isBuiltinWithName("str")
+    assertThat(typeCheckerContext.typeChecker().isBuiltinWithName("int").or(
+      typeCheckerContext.typeChecker().isBuiltinWithName("str")
     ).check(intLiteralType)).isEqualTo(TriBool.TRUE);
 
-    assertThat(typeChecker.typeCheckBuilder().isBuiltinWithName("int").or(
-      typeChecker.typeCheckBuilder().hasMember("__abs__")
+    assertThat(typeCheckerContext.typeChecker().isBuiltinWithName("int").or(
+      typeCheckerContext.typeChecker().hasMember("__abs__")
     ).check(intLiteralType)).isEqualTo(TriBool.TRUE);
   }
 
@@ -123,11 +123,11 @@ class TypeCheckerTest {
     NumericLiteral intLiteral = (NumericLiteral) TreeUtils.firstChild(fileInput, t -> t.is(Tree.Kind.NUMERIC_LITERAL)).get();
     ObjectType intLiteralType = (ObjectType) intLiteral.typeV2();
 
-    TypeCheckBuilder isIntAndStr = typeChecker.typeCheckBuilder().isBuiltinWithName("int").isBuiltinWithName("str");
-    TypeCheckBuilder isStrAndInt = typeChecker.typeCheckBuilder().isBuiltinWithName("str").isBuiltinWithName("int");
-    TypeCheckBuilder isIntAndHasMemberAbs = typeChecker.typeCheckBuilder().isBuiltinWithName("int").hasMember("__abs__");
-    TypeCheckBuilder hasMemberFloatAndEq = typeChecker.typeCheckBuilder().hasMember("__eq__").hasMember("__float__");
-    TypeCheckBuilder isUnknownAndHasMemberFloat = typeChecker.typeCheckBuilder().isBuiltinWithName("UNKNOWN_BUILTIN_NAME").hasMember("__float__");
+    TypeChecker isIntAndStr = typeCheckerContext.typeChecker().isBuiltinWithName("int").isBuiltinWithName("str");
+    TypeChecker isStrAndInt = typeCheckerContext.typeChecker().isBuiltinWithName("str").isBuiltinWithName("int");
+    TypeChecker isIntAndHasMemberAbs = typeCheckerContext.typeChecker().isBuiltinWithName("int").hasMember("__abs__");
+    TypeChecker hasMemberFloatAndEq = typeCheckerContext.typeChecker().hasMember("__eq__").hasMember("__float__");
+    TypeChecker isUnknownAndHasMemberFloat = typeCheckerContext.typeChecker().isBuiltinWithName("UNKNOWN_BUILTIN_NAME").hasMember("__float__");
 
     assertThat(isIntAndStr.check(intLiteralType)).isEqualTo(TriBool.FALSE);
     assertThat(isStrAndInt.check(intLiteralType)).isEqualTo(TriBool.FALSE);
@@ -139,10 +139,10 @@ class TypeCheckerTest {
     assertThat(isUnknownAndHasMemberFloat.and(isIntAndHasMemberAbs).check(intLiteralType)).isEqualTo(TriBool.UNKNOWN);
     assertThat(isIntAndStr.and(isStrAndInt).check(intLiteralType)).isEqualTo(TriBool.FALSE);
 
-    TypeCheckBuilder trueOr1 = isStrAndInt.or(isIntAndHasMemberAbs);
-    TypeCheckBuilder trueOr2 = isIntAndHasMemberAbs.or(isStrAndInt);
-    TypeCheckBuilder trueOr3 = isIntAndHasMemberAbs.or(isUnknownAndHasMemberFloat);
-    TypeCheckBuilder trueOr4 = isUnknownAndHasMemberFloat.or(isIntAndHasMemberAbs);
+    TypeChecker trueOr1 = isStrAndInt.or(isIntAndHasMemberAbs);
+    TypeChecker trueOr2 = isIntAndHasMemberAbs.or(isStrAndInt);
+    TypeChecker trueOr3 = isIntAndHasMemberAbs.or(isUnknownAndHasMemberFloat);
+    TypeChecker trueOr4 = isUnknownAndHasMemberFloat.or(isIntAndHasMemberAbs);
 
     assertThat(trueOr1.check(intLiteralType)).isEqualTo(TriBool.TRUE);
     assertThat(trueOr2.check(intLiteralType)).isEqualTo(TriBool.TRUE);
